Fix bug realpath + code refactoring

+++ url_parser.c
@@ -32,8 +32,8 @@ struct parsed_url *parse_url(const char *url) {
     struct parsed_url *purl;
     const char *tmpstr;
     const char *curstr;
-    long len;
-    int i;
+    size_t len;
+    unsigned int i;
     int userpass_flag;
     int bracket_flag;
